Remove the PR store. (#251)
* Add PR trigger for freestyle (and pipeline) projects as well

* Fixing parameter ordering to be consistent, and rename some fields and parameters to be clearer and consistent

* Fix up upgrades and default value handling

* Refactor Webhook event code, remove usage of deprecated methods. More general tidiup.

* Minor code review improvements

* Fix javadoc a bit

* Tiny javadoc tweaks

* Remove the PR store.

We do not have the capacity to properly review and test the PullRequestStore right now. Since we plan to release a version of "Build on PR" that does not strictly require it, the store is removed. There is a branch prStoreTrait that preserves this code so we do not loose it and can thus test it properly before release.

* Add support for Source Branch updated event (#252)